refactor after detekt
