+++ app/src/main/java/com/innoprog/android/feature/profile/profiledetails/di/ProfileModule.kt
@@ -1,14 +1,14 @@
 package com.innoprog.android.feature.profile.profiledetails.di
 
-import com.innoprog.android.feature.profile.profiledetails.presentation.ProfileViewModel
 import androidx.lifecycle.ViewModel
 import com.innoprog.android.di.ViewModelKey
 import com.innoprog.android.feature.profile.profiledetails.data.impl.ProfileInfoRepoImpl
 import com.innoprog.android.feature.profile.profiledetails.data.network.NetworkClient
 import com.innoprog.android.feature.profile.profiledetails.data.network.RetrofitNetworkClient
 import com.innoprog.android.feature.profile.profiledetails.domain.GetProfileUseCase
-import com.innoprog.android.feature.profile.profiledetails.domain.impl.GetProfileUseCaseImpl
 import com.innoprog.android.feature.profile.profiledetails.domain.ProfileInfoRepo
+import com.innoprog.android.feature.profile.profiledetails.domain.impl.GetProfileUseCaseImpl
+import com.innoprog.android.feature.profile.profiledetails.presentation.ProfileViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
@@ -29,6 +29,4 @@ interface ProfileModule {
 
     @Binds
     fun provideNetworkClient(impl: RetrofitNetworkClient): NetworkClient
-
 }
-
