+++ app/src/main/java/com/innoprog/android/feature/profile/profiledetails/presentation/ProfileScreenState.kt
@@ -4,7 +4,6 @@ import com.innoprog.android.feature.profile.profiledetails.domain.models.Profile
 import com.innoprog.android.util.ErrorType
 
 sealed interface ProfileScreenState {
-    data object Loading : ProfileScreenState
 
     data class Content(val profileInfo: Profile) : ProfileScreenState
