+++ app/src/main/java/com/innoprog/android/feature/profile/profiledetails/presentation/ProfileViewModel.kt
@@ -5,15 +5,11 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
 import com.innoprog.android.base.BaseViewModel
 import com.innoprog.android.feature.profile.profiledetails.domain.GetProfileUseCase
-import com.innoprog.android.feature.profile.profiledetails.domain.ProfileInfoRepo
 import com.innoprog.android.feature.profile.profiledetails.domain.models.Profile
+import com.innoprog.android.util.ErrorType
 import com.innoprog.android.util.Resource
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
-import java.io.IOException
 import javax.inject.Inject
 
 class ProfileViewModel @Inject constructor(private val getProfileUseCase: GetProfileUseCase) :
@@ -27,7 +23,8 @@ class ProfileViewModel @Inject constructor(private val getProfileUseCase: GetPro
         if (profile != null) {
             _uiState.postValue(ProfileScreenState.Content(profile as Profile))
         } else {
-            _uiState.value = ProfileScreenState.Loading
+            _uiState.value = ProfileScreenState.Error(ErrorType.NOT_FOUND)
+
             viewModelScope.launch(Dispatchers.IO) {
                 getProfileUseCase.getProfile().collect { response ->
                     when (response) {
@@ -35,9 +32,8 @@ class ProfileViewModel @Inject constructor(private val getProfileUseCase: GetPro
                             _uiState.postValue(ProfileScreenState.Content(response.data))
                             profile = response.data
                         }
-
                         is Resource.Error -> {
-                                _uiState.postValue(ProfileScreenState.Error(response.errorType))
+                            _uiState.postValue(ProfileScreenState.Error(response.errorType))
                         }
                     }
                 }
