+++ app/src/main/java/com/innoprog/android/util/ErrorType.kt
